+++ src/styles/themes/next-gen/codeView/codeView.ts
@@ -146,17 +146,20 @@ const lineNo = {
 };
 
 export const copyButton = {
+  path: {
+    fill: 'text.primary',
+  },
   default: {
     mx: 'sm',
     path: {
-      fill: 'font.reverse',
+      fill: 'white',
     },
     border: 'none !important',
   },
   light: {
     mx: 'sm',
     path: {
-      fill: '#23282f',
+      fill: 'black',
     },
     border: 'none !important',
   },